Add files via upload
Not fully working, but this is my first commit, so I thought I might as well toss this in here. 

Code reads in .txt file identical to the one outputted by STK's Access Report generator, and outputs an array with the number of seconds since mission start. Still a lot of optimizing that needs to be done.